+++ src/api/api_models/common/SignedProposalMessage.ts
@@ -1,0 +1,4 @@
+export default interface SignedProposalMessage {
+    unsignedProposal: string;
+    signature: string;
+}