+++ src/components/course/list/common/CourseList.vue
@@ -78,8 +78,8 @@
                 }
                 courses.value = genericResponseHandler.handleResponse(response);
                 const lecturerIds = new Set(courses.value.map((course) => course.lecturerId));
-                const resp = await userManagement.getLecturers(...lecturerIds);
-                lecturers.value = genericResponseHandler.handleResponse(resp);
+                const resp = await userManagement.getUsers(Role.LECTURER, [...lecturerIds]);
+                lecturers.value = genericResponseHandler.handleResponse(resp) as Lecturer[];
                 busy.value = false;
             }
 
@@ -103,8 +103,8 @@
                     let filter = props.filter.toLowerCase();
                     filteredCourses = filteredCourses.filter(
                         (e) =>
-                            e.courseName.toLowerCase().includes(filter) ||
-                            e.courseDescription.toLowerCase().includes(filter) ||
+                            e.courseName.toLowerCase().replace(/\s/g, "").includes(filter.replace(/\s/g, "")) ||
+                            e.courseDescription.toLowerCase().replace(/\s/g, "").includes(filter.replace(/\s/g, "")) ||
                             e.courseId.toLowerCase().includes(filter) ||
                             e.lecturerId.toLowerCase().includes(filter)
                     );