+++ src/components/navigation/navbar/desktop/student/Navbar.vue
@@ -6,6 +6,9 @@
         <li>
             <course-menu id="nav_desktop_student_menu_courses" class="navbar-menu" />
         </li>
+        <li>
+            <study-menu id="nav_desktop_student_menu_study" class="navbar-menu" />
+        </li>
         <template #right>
             <profile-menu id="menu_profile" class="pt-4" />
         </template>
@@ -15,6 +18,7 @@
 <script lang="ts">
     import BaseNavbar from "../common/BaseNavbar.vue";
     import CourseMenu from "./courses/CourseMenu.vue";
+    import StudyMenu from "./study/StudyMenu.vue";
     import ProfileMenu from "../common/profile/ProfileMenu.vue";
 
     export default {
@@ -23,6 +27,7 @@
             BaseNavbar,
             CourseMenu,
             ProfileMenu,
+            StudyMenu,
         },
     };
 </script>