+++ src/use/helpers/GenericResponseHandler.ts
@@ -16,7 +16,9 @@ export default class GenericResponseHandler implements ResponseHandler<boolean> 
             return response.returnValue;
         }
         switch (response.statusCode) {
-            case 400: {
+            case 400:
+            case 500:
+            case 503: {
                 showAPIToast(response.statusCode);
                 return response.returnValue;
             }
@@ -25,15 +27,11 @@ export default class GenericResponseHandler implements ResponseHandler<boolean> 
                 return response.returnValue;
             }
             case 403: {
-                alert("You do not have the neccessary user rights for this action!");
+                showAPIToast(response.statusCode, this.dataType);
                 return response.returnValue;
             }
             case 404: {
                 showAPIToast(response.statusCode, this.dataType);
-                return response.returnValue;
-            }
-            case 500: {
-                showAPIToast(response.statusCode);
                 return response.returnValue;
             }
             case 200: {