+++ src/use/helpers/Toasts.ts
@@ -15,8 +15,14 @@ export function showAPIToast(code: number, msg?: string) {
             //toast.error(markRaw(ReportBug));
             toast.error("Something went wrong on our side. Please consider reporting this bug.");
             break;
+        case 403:
+            toast.error("You are not authorized for " + msg + "!");
+            break;
         case 404:
             toast.error("Could not find " + msg + ".");
+            break;
+        case 503:
+            toast.error("Service unavailable. Please consider reporting this.");
             break;
         default:
             toast.error("Something went wrong. Please try again later.");