+++ src/use/helpers/Versions.ts
@@ -1,29 +1,46 @@
+import AdmissionManagement from "@/api/AdmissionManagement";
 import AuthenticationManagement from "@/api/AuthenticationManagement";
 import CertificateManagement from "@/api/CertificateManagement";
 import CourseManagement from "@/api/CourseManagement";
 import ExaminationRegulationManagement from "@/api/ExaminationRegulationManagement";
 import MatriculationManagement from "@/api/MatriculationManagement";
+import ReportManagement from "@/api/ReportManagement";
 import UserManagement from "@/api/UserManagement";
-import { useStore } from "../store/store";
 
 export interface version {
     name: String;
     version: String;
     link: String;
+    hlVersions?: {
+        hlfApiVersion: string;
+        chaincodeVersion: string;
+    };
 }
 
 export async function getVersions(): Promise<version[]> {
-    let versions: version[] = [];
-    let store = useStore();
+    const versions: version[] = [];
 
-    let frontEndVersion = "v" + process.env.VUE_APP_VERSION;
+    const frontEndVersion = "v" + process.env.VUE_APP_VERSION;
 
-    let authenticationManagementVersion = await AuthenticationManagement.getVersion();
-    let courseManagementVersion = await CourseManagement.getVersion();
-    let userManagementVersion = await UserManagement.getVersion();
-    let matriculationManagementVersion = await MatriculationManagement.getVersion();
-    let certificateManagementVersion = await CertificateManagement.getVersion();
-    let examinationRegulationManagementVersion = await ExaminationRegulationManagement.getVersion();
+    const authenticationManagementVersion = await AuthenticationManagement.getVersion();
+
+    const courseManagementVersion = await CourseManagement.getVersion();
+
+    const userManagementVersion = await UserManagement.getVersion();
+
+    const matriculationManagementVersion = await MatriculationManagement.getVersion();
+    const hlfMatriculationVersion = await MatriculationManagement.getHyperledgerVersion();
+
+    const certificateManagementVersion = await CertificateManagement.getVersion();
+    const hlfCertificateVersion = await CertificateManagement.getHyperledgerVersion();
+
+    const examinationRegulationManagementVersion = await ExaminationRegulationManagement.getVersion();
+    const hlfExamRegVersion = await ExaminationRegulationManagement.getHyperledgerVersion();
+
+    const reportManagementVersion = await ReportManagement.getVersion();
+
+    const admissionsManagementVersion = await AdmissionManagement.getVersion();
+    const hlfAdmissionVersion = await AdmissionManagement.getHyperledgerVersion();
 
     versions.push({
         name: "Frontend",
@@ -65,6 +82,7 @@ export async function getVersions(): Promise<version[]> {
             "matriculation-" +
             matriculationManagementVersion +
             "/product_code/matriculation_service/CHANGELOG.md",
+        hlVersions: hlfMatriculationVersion,
     });
     versions.push({
         name: "Certificate Management",
@@ -74,6 +92,7 @@ export async function getVersions(): Promise<version[]> {
             "certificate-" +
             certificateManagementVersion +
             "/product_code/certificate_service/CHANGELOG.md",
+        hlVersions: hlfCertificateVersion,
     });
     versions.push({
         name: "Examination Regulations Management",
@@ -83,6 +102,27 @@ export async function getVersions(): Promise<version[]> {
             "examreg-" +
             examinationRegulationManagementVersion +
             "/product_code/examreg_service/CHANGELOG.md",
+        hlVersions: hlfExamRegVersion,
+    });
+    versions.push({
+        name: "Report Management",
+        version: reportManagementVersion,
+        link:
+            "https://github.com/upb-uc4/University-Credits-4.0/blob/" +
+            "report-" +
+            reportManagementVersion +
+            "/product_code/report_service/CHANGELOG.md",
+    });
+
+    versions.push({
+        name: "Admissions Management",
+        version: admissionsManagementVersion,
+        link:
+            "https://github.com/upb-uc4/University-Credits-4.0/blob/" +
+            "admission-" +
+            admissionsManagementVersion +
+            "/product_code/admission_service/CHANGELOG.md",
+        hlVersions: hlfAdmissionVersion,
     });
 
     return versions;