+++ src/use/router/index.ts
@@ -17,7 +17,9 @@ const CourseForm = () => import("@/views/shared/EditCreateCourseForm.vue");
 const AccountForm = () => import("@/views/admin/EditCreateAccountForm.vue");
 const PrivateProfile = () => import("@/views/common/PrivateProfile.vue");
 const PublicProfile = () => import("@/views/common/PublicProfile.vue");
+const ExamRegForm = () => import("@/views/admin/CreateExamRegForm.vue");
 const Settings = () => import("@/views/common/Settings.vue");
+const StudentImmatricultaion = () => import("@/components/study/Immatriculation.vue");
 
 const routerHistory = createWebHistory(process.env.BASE_URL);
 const suffix: string = " | UC4";
@@ -45,7 +47,7 @@ const router = createRouter({
         {
             path: "/course-catalog",
             name: "courseCatalog",
-            props: { showAllCourses: true },
+            props: { showAllCourses: true, isCourseCatalogue: true },
             component: AllCourseView,
             meta: {
                 title: "Course Catalog" + suffix,
@@ -68,6 +70,17 @@ const router = createRouter({
                 roles: ["Student"],
             },
         },
+
+        {
+            path: "/immatriculation",
+            name: "student.immatriculation",
+            component: StudentImmatricultaion,
+            meta: {
+                title: "My Immatriculation" + suffix,
+                roles: ["Student"],
+            },
+        },
+
         {
             path: "/courses/my-courses",
             name: "student.my-courses",
@@ -151,6 +164,15 @@ const router = createRouter({
             meta: {
                 title: "Course Editing" + suffix,
                 roles: ["Lecturer", "Admin"],
+            },
+        },
+        {
+            path: "/create-exam-reg",
+            name: "examRegForm",
+            component: ExamRegForm,
+            meta: {
+                title: "Examination Regulation Creation" + suffix,
+                roles: ["Admin"],
             },
         },
         {