+++ src/use/store/getters.ts
@@ -4,6 +4,7 @@ import Lecturer from "@/api/api_models/user_management/Lecturer";
 import Student from "@/api/api_models/user_management/Student";
 import AuthenticationManagement from "@/api/AuthenticationManagement";
 import CertificateManagement from "@/api/CertificateManagement";
+import { useToast } from "@/toast";
 import GenericResponseHandler from "@/use/helpers/GenericResponseHandler";
 import { GetterTree } from "vuex";
 import { ActionTypes } from "./action-types";
@@ -15,7 +16,7 @@ import { useStore } from "./store";
 export type Getters = {
     user(state: State): Promise<Student | Lecturer | Admin>;
     loggedIn(state: State): boolean;
-    privateKey(state: State): Promise<CryptoKey>;
+    privateKey(state: State): () => Promise<CryptoKey>;
     certificate(state: State): () => Promise<Certificate>;
     hasCertificate(state: State): Promise<boolean>;
 };
@@ -41,12 +42,13 @@ export const getters: GetterTree<State, State> & Getters = {
         const store = useStore();
         return (await store.getters.user).role;
     },
-    privateKey: async (state) => {
+    privateKey: (state) => async () => {
         if (!("type" in state.privateKey)) {
             const store = useStore();
             const certManagement = new CertificateManagement();
 
             const keyResponse = await certManagement.getEncryptedPrivateKey((await store.getters.user).username);
+
             const handler = new GenericResponseHandler("certificate");
             const encryptedPrivateKey = handler.handleResponse(keyResponse);
 
@@ -55,10 +57,32 @@ export const getters: GetterTree<State, State> & Getters = {
                 const privateKey: CryptoKey = await state.decryptPrivateKeyModal(encryptedPrivateKey);
 
                 if (privateKey.type == undefined) {
-                    Promise.reject("Could not decrypt private key");
+                    return Promise.reject("Could not decrypt private key");
                 }
 
                 store.commit(MutationTypes.SET_PRIVATE_KEY, privateKey);
+            } else {
+                // no key saved at lagom
+                if (state.certificate.certificate) {
+                    const toast = useToast();
+                    toast.error(
+                        "Something went wrong. this user does not have an encrypted private key saved but has a certificate. Please report this."
+                    );
+                } else {
+                    // create certificate
+                    const certificate = await store.dispatch(ActionTypes.CREATE_CERTIFICATE, undefined).catch((reason) => {
+                        store.commit(MutationTypes.SET_HAS_CERTIFICATE, false);
+                        return undefined;
+                    });
+
+                    if (!certificate) return Promise.reject("Could not create certificate.");
+
+                    const keyResponse = await certManagement.getEncryptedPrivateKey((await store.getters.user).username);
+                    const encryptedPrivateKey = handler.handleResponse(keyResponse);
+                    const privateKey: CryptoKey = await state.decryptPrivateKeyModal(encryptedPrivateKey);
+
+                    store.commit(MutationTypes.SET_PRIVATE_KEY, privateKey);
+                }
             }
         }
         return state.privateKey;
@@ -72,7 +96,7 @@ export const getters: GetterTree<State, State> & Getters = {
             if (certificateResponse.statusCode == 404) {
                 return await store.dispatch(ActionTypes.CREATE_CERTIFICATE, undefined).catch((reason) => {
                     store.commit(MutationTypes.SET_HAS_CERTIFICATE, false);
-                    return { certificate: "" };
+                    return Promise.reject("Could not create certificate.");
                 });
             } else {
                 store.commit(MutationTypes.SET_CERTIFICATE, certificateResponse.returnValue);