+++ src/views/common/Settings.vue
@@ -10,6 +10,10 @@
         <div>
             <security-section />
             <certificate-section />
+            <request-data-section />
+            <div class="mt-48">
+                <delete-account-section />
+            </div>
         </div>
     </div>
 </template>
@@ -20,12 +24,16 @@
     import { Role } from "@/entities/Role";
     import { checkPrivilege } from "@/use/helpers/PermissionHelper";
     import CertificateSection from "@/components/settings/CertificateSection.vue";
+    import DeleteAccountSection from "@/components/settings/DeleteAccountSection.vue";
+    import RequestDataSection from "@/components/settings/RequestDataSection.vue";
 
     export default {
         name: "Settings",
         components: {
             SecuritySection,
             CertificateSection,
+            DeleteAccountSection,
+            RequestDataSection,
         },
 
         async beforeRouteEnter(_from: any, _to: any, next: any) {