+++ src/views/lecturer/PrivateProfile.vue
@@ -20,31 +20,6 @@
             <address-section v-model:user="lecturer" />
 
             <research-section v-model:user="lecturer" />
-
-            <section class="border-t-2 py-8 border-gray-400">
-                <div class="lg:flex">
-                    <div class="w-full lg:w-1/3 lg:block mr-12 flex flex-col mb-4">
-                        <label class="block text-gray-700 text-lg font-medium mb-2">Notifications</label>
-                        <label class="block text-gray-600"> Control when and how to stay updated about recent activities. </label>
-                    </div>
-                    <div class="w-full lg:w-2/3">
-                        <div class="mb-6 flex items-start">
-                            <input id="subscribeCourses" class="mt-2 w-6 h-6 form-checkbox text-blue-600" type="checkbox" checked />
-                            <div class="ml-6 flex flex-col">
-                                <label class="text-gray-700 text-lg font-medium">Course Activity</label>
-                                <label class="text-gray-600">Get important notifications about the courses you are enrolled in</label>
-                            </div>
-                        </div>
-                        <div class="mb-6 flex items-start">
-                            <input id="otherCheckbox" class="mt-2 w-6 h-6 form-checkbox text-blue-600" type="checkbox" checked />
-                            <div class="ml-6 flex flex-col">
-                                <label class="text-gray-700 text-lg font-medium">Some Other Stuff</label>
-                                <label class="text-gray-600">This was just an idea. Feel free to delete it.</label>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </section>
         </div>
     </div>
 </template>